Refactor common functions into common utility lib
